Rework file structure and move App component to new 'components' folder. Separate navbar into its component
