--- src/App.css
@@ -1,8 +1,0 @@
-.app {
-    --color-dark: #2B2B2B;
-    --color-light: #FCFCFC;
-    --color-primary: #EB5E28;
-    --color-secondary: #337CA0;
-    --color-tertiary: #BDC667;
-}
-
