+++ src/components/App.css
@@ -1,0 +1,8 @@
+.app {
+  --color-dark: #2B2B2B;
+  --color-light: #FCFCFC;
+  --color-primary: #F9DC5C;
+  --color-secondary: #33658A;
+  --color-accent: #ED254E;
+}
+
