+++ src/components/navbar/Navbar.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import {Target} from 'react-feather';
+
+import './Navbar.css';
+
+export const Navbar = () => {
+  return (
+    <header className="navbar">
+      <Target className="navbar-icon" size="2.8em" />
+      <h1 className="navbar-title"> El Baratón </h1>
+  </header>
+);
+}
